+++ overlays/terraform.nix
@@ -6,13 +6,13 @@ self: super: {
       else
         plugin) {
           onepassword = super.terraform-providers.mkProvider
-            (let version = "1.4.0";
+            (let version = "2.1.0";
             in {
               inherit version;
               owner = "1Password";
               repo = "terraform-provider-onepassword";
               rev = "v${version}";
-              hash = "sha256-+zusrzMZqoPsSIv2Dh2IEkFd17HVDrVj2wp5th4rRjk=";
+              hash = "sha256-rdS9Udzfc/U7E4CIyySnntOCVBBZL0/GuAiVCI5uMrc=";
               vendorHash = null;
               provider-source-address =
                 "registry.terraform.io/1Password/onepassword";